Adding single/multi use context wrapper
Adding single/multi use context wrapper to maintain object contexts for lazy loading when being removed from the write-context. This allows the DB connection to be maintained and preserves navigation properties but hides the underlying context after the write context is closed so the next context request will return a new context which prevents objects from the previous context from being accidentally updated which is what the former detatch used to accomplish, however this unset the navigation properties.

The default scope for a DbContext should now be set to Transient and the IDbContext interface should be used for Persistence Repositories which allows the Multi-Use context wrapper to manage the internal context lifecycle on its own and only request a new DbContext implementation when the previous write-contexts are closed.

The multi-use db context wrapper should be set to scoped use in place of the former scoped DbContext